refactor: resume editor
